PostgreSQL/requirements.txt
azure-functions
azure-functions-durable
azure.cli.core
pyodbc
pymongo
numpyencoder
mysql.connector
datetime
psycopg2
psycopg2-binary